Bug 1787526 - Part 7: Implement semispace nursery r=sfink
Implement semispace nursery by promoting some cells to a second nursery
generation. Since we've swapped semispaces and are evicting from-space to
to-sapce, promotion works like normal allocation. In the same way, swapping
store buffers makes post barriers also work like normal.

While tracing store buffers, we may need to re-add the store buffer entry. The
TenuringTrace has a flag that is set if any allocation is made into the next
generation which is used to check whether this is necessary.

Whether to tenure a particular nursery cell is based on its position in the
nursery. Cells that were promoted in the last collection are tenured.

Sometimes we tenure everything and totally empty the nursery. This is decided
based on the nursery collection reason. The GC uses the EVICT_NURSERY reason
to make sure everything is tenured during major GC.

Differential Revision: https://phabricator.services.mozilla.com/D196440

UltraBlame original commit: 855def3146ee904eaa064218609ce5cfa16e48ba