Bug 1883336 - Pass two ints instead of one as shader input. r=gw
It's convenient for shaders that need to fetch data from two different places and we have a spare slot in the quad header. It's not strictly necessary if we want the spare slot for something else, since we could use the integer gpu buffer as an indirection for when we need more than one address (although that's a tad convoluted and inefficient).

Differential Revision: https://phabricator.services.mozilla.com/D204386

UltraBlame original commit: 43d89580ba9c295f67bc3df94ba1aac23b435b63