Bug 1880814 - Logs for libvpx configuration on various platforms r=glandium
This patch generates log files during the configuration of libvpx on
variouse platforms. These logs can be used for manual verification of
the correctness of settings. In this particular case, the log file for
win/aarch64 reveals the sve feature is disabled on win/aarch64.

The logs will be excluded to the repo, by .gitignore and .hgignore
settings.

Differential Revision: https://phabricator.services.mozilla.com/D204799

UltraBlame original commit: 0da84f0cfef667fe81244d4085285508bbf7fbce