Bug 1883457 - Part 2: Use be_memory_inline_jit_restrict_* APIs for JIT on iOS, r=jandem
This change enables the JS_USE_APPLE_FAST_WX option for iOS, but uses the
BrowserEngineKit APIs rather than the pthread_jit APIs which are available on
macOS. It is unclear to me if there are other differences with these APIs which
would need to be handled beyond what has been written so far, as this is just a
naive substitution.

Given these functions are explicitly "inline", it may be desirable at some
point to refactor the code to allow the calls to be made inline within
ProcessExecutableMemory.h, however that was left out-of-scope.

Differential Revision: https://phabricator.services.mozilla.com/D203499

UltraBlame original commit: 2dbe5fbe8f16056095e9003a219d1bace280edce