Bug 1806400 - Activate the target element in ProcessSingleTap() on the first tap if double-tap didn't happen. r=dlrobertson,botond
On environments where double-tap-to-zoom is allowed, we don't immediately send
a single tap event when a touch-end happens, we later send a single tap event
when we are sure that the touch-end is not followed by double-tapping. Thus
when ActiveElementManager got the touch-end notification from an APZC, we need
to differentiate the situation to activate an element when ActiveElementManager
finally got a single tap event. This change introduces a tri-state to differentiate
the situation.

`helper_bug1806400-3.html` in this change is a subset of
`helper_touch_synthesized_mouseevents.html?scrollable=true`.

`helper_bug1806400-4.html` is anocher variant of `helper_bug1806400-2.html`

Differential Revision: https://phabricator.services.mozilla.com/D202384

UltraBlame original commit: 637b67cd5a714e0d7addb2668f8672d216ee5206