Bug 1886259 [wpt PR 45201] - Limit use of calc-size() within other expression contexts., a=testonly
Automatic update from web-platform-tests
Limit use of calc-size() within other expression contexts.

This restricts the use of calc-size() to the toplevel expression and to
the basis of another calc-size() expression.  (It should eventually also
be allowed in mix() when we support that.)  It implements the
restriction in the same way that the spec describes it: based on the
idea that its type is not a length.

Bug: 313072
Change-Id: I157a16fee67674511f22f7d604c259c2230ab79c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5367589
Reviewed-by: Daniil Sakhapov <sakhapovchromium.org>
Commit-Queue: David Baron <dbaronchromium.org>
Cr-Commit-Position: refs/heads/main{#1276371}

--

wpt-commits: 8f3aba888753b219abda53996fb523ee2e30e8b1
wpt-pr: 45201

UltraBlame original commit: e5f8d9396cfc602f42f085dedd10d4a3312efe90