Bug 1886252 - Modernize some old AOM xpinstall tests and tweak them to compute the expected xpi file hash. r=willdurand
There are some other of these test files still using a pattern we don't use in any new tests, but for now I only took care
of the ones that would be failing with a new amosigned.xpi file replacing the old one.
Based on the final version of these ones we can then file a separate followup to tweak the other if we think that would
be valuable.

Differential Revision: https://phabricator.services.mozilla.com/D205768

UltraBlame original commit: 31d8e5781020850aa707b20de6a90c33b5c0ba6d