Bug 1885571 - Quad radial gradient shader. r=gw
For now we skip the quad path if the gradient is tiled. The plan is to add a repeating quad pattern that can repeat the result of a render task and get all quad patterns to use that.

Differential Revision: https://phabricator.services.mozilla.com/D204774

UltraBlame original commit: 09548e11b44f96ff9fe8f48c191a5e10f84899b1