Bug 1887954 - Make `nsIFrame::GetLastLeaf()` never returns a frame in native anonymous subtrees under the given frame r=emilio
It does not check whether the child frame is in a native anonymous subtree
only when every first child at digging the frame tree.  This patch rewrites
the complicated loop with 2 for-loops and make it always check whether the
frame is for a native anonymous subtree root or not before updating the result.

Differential Revision: https://phabricator.services.mozilla.com/D206283

UltraBlame original commit: 6d76b71d89f3d3218632b68cecf9c7295b4fb41b