Bug 1689705 - Remove marquee events. r=smaug
After bug 306344 this is the last remaining incompatibility with
<marquee>. I think we should be able to drop it. Use counters for
"finish" are relatively high (>0.1%), but I'm pretty sure that's because
that event name is shared with web animations. "bounce" is a more
realistic measure, and that's 0.003%, which is more expected.

Differential Revision: https://phabricator.services.mozilla.com/D206387

UltraBlame original commit: 9455e560433a94b062454fc7abe5d8f01f693743