Bug 1889166. Invalidate images in popups on macOS which aren't drawn with webrender. r=layout-reviewers,emilio
The problem is that in InvalidateImages in ImageLoader.cpp

https://searchfox.org/mozilla-central/rev/e1e4a33e82ee1d278df238cf0896b7358a4bc359/layout/style/ImageLoader.cpp#517

we do nothing to invalidate images that are in popups that aren't drawn by webrender. This is controlled via ShouldUseOffMainThreadCompositing here

https://searchfox.org/mozilla-central/rev/5df192b9edee971a1ab06ff9ef688d92c21a92c1/widget/cocoa/nsChildView.mm#1245

So all popups on macOS are painted using the fallback renderer. So InvalidateImages will do absolutely nothing in this case because there is no saved webrender user data, and the aForcePaint flag is false (because this is a frame update notification, not a frame complete/first frame notification).

Differential Revision: https://phabricator.services.mozilla.com/D206569

UltraBlame original commit: 665c1645364137aa63f4d4240bfd1ce313f200fc