Bug 1890021 - Reimplement loginToVanillaObject in content JS. r=credential-management-reviewers,sgalich
This code is being run as content JS because this is in a plain Mochitest, but
it is calling into the chrome JS function LoginHelper.loginToVanillaObject().
That function returns a chrome JS object, so it ends up in a cross-compartment
wrapper (CCW). The code then attempts to send the CCW over JS actor IPC, via
JSActor::SendQuery. That method attempts to structured clone the CCW by calling
into nsFrameMessageManager::GetParamsForMessage(), but structured cloning
can't deal with a CCW to a plain object, so it falls back to the stringify
JSON method. JS_Stringify apparently can deal with a CCW, so it produces
a JSON string which gets turned into a new JS value, which can be successfully
structured cloned.

The goal of this patch is to avoid relying on that JSON fallback for structured
clone. It does this by reimplementing loginToVanillaObject in content JS, so the
resulting object is content JS, so we avoid the CCW.

Differential Revision: https://phabricator.services.mozilla.com/D206809

UltraBlame original commit: 1564b30ff7230370f1d96bdfc88b528c6dbf1f37