Bug 1882704 [wpt PR 44849] - [line-clamp] Hide out of flow positioned boxes after the clamp point, a=testonly
Automatic update from web-platform-tests
[line-clamp] Hide out of flow positioned boxes after the clamp point

This patch makes out of flow positioned boxes that appear in the
layout tree after the clamp point (that is, in tree order after the
end of the last shown line) hidden for paint. This happens depending
entirely on the layout tree order, regardless of where that box is
positioned.

This is achieved by storing a flag in the various `*OofPositionedNode`
classes to mark the node as hidden for paint. This flag is then
checked when creating the constraint space for the layout of the box.

Bug: 40336192
Change-Id: Ia42751b2fa7e3fee89648e26600035241bda6299
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5331654
Commit-Queue: Andreu Botella <abotellaigalia.com>
Reviewed-by: Morten Stenshorne <mstenshochromium.org>
Cr-Commit-Position: refs/heads/main{#1283041}

--

wpt-commits: eef6b7e5a955ced8522e11b2c6ff6537759ea6aa
wpt-pr: 44849

UltraBlame original commit: 4a13440c7e157e803541bcf290068c4328e71197