Bug 1849160 - Part 2: Add Accessible::FindAncestorIf, r=Jamie
This revision adds a utility function to find the nearest non-generic ancestor
for which a given predicate returns AncestorSearchOption::Found. The goal of
this function is to enable searching through ancestors, stopping when
appropriate. This is useful in a follow-up patch for finding requisite ancestors
of child role accessibility nodes.

Differential Revision: https://phabricator.services.mozilla.com/D205672

UltraBlame original commit: c8e9ae6999813fdf693cd4083df70c0d72c8ce64