Bug 1888429 - Part 4: Make the internal thread pool dispatch tasks to free threads in a fixed order r=jandem
This changes the shell's internal thread pool to dispatch tasks to helper threads in
a fixed order rather than by threads picking up each task depending on which is
the first to get scheduled. In other words, if there is only ever at most one
task in the system then all tasks will be run on the first pool thread; if
there are at most two then the only the first two threads will be used.

This results in a noticable improvement on the splay benchmark on my linux
system, probably due to cache effects. This patch only affects the shell but it
indicates that it may be preferable for TaskController to work in a similar
way.

This removes JS::DispatchReason which is no longer needed.

Differential Revision: https://phabricator.services.mozilla.com/D206764

UltraBlame original commit: 4c5efaca40578ee5b3b73d32fe7f6888e2c8c65e