Bug 1885571 - Quad radial gradient shader. r=gw
For now we skip the quad path if the gradient is tiled. The plan is to add a repeating quad pattern that can repeat the result of a render task and get all quad patterns to use that.

Differential Revision: https://phabricator.services.mozilla.com/D204774

UltraBlame original commit: cd0d592ecdac96aae396b4979c9d40bf61c318c1