Bug 1884256 - Forbid GIFFT mirroring of non-ping-lifetime metrics r=janerik,rpl
Most `lifetime: application` metrics are granted the new no_lint:
"GIFFT_NON_PING_LIFETIME" so they pass. And since it is possible that people
might wish to do this deliberately and not only be mistake, the no_lint isn't
exclusive-via-allowlist to just existing metrics.

The _distribution metrics in extension are changed to `lifetime: ping` since
that appeared to be a... not exactly a mistake, but perhaps a suboptimal choice,
in instrumentation. This will later be enforced by glean_parser-wide lint in
bug 1886810 (forbid non-ping-lifetime _distributions (regardless of mirror)).

Differential Revision: https://phabricator.services.mozilla.com/D207639

UltraBlame original commit: 1f57e3ccfcd7a1d6a3d56779279dadf48a9f7428